Removed leftover utils
