+++ jobs/dtd-metrics/src/metrics/tenant-onboarding-trend.metric.ts
@@ -1,6 +1,7 @@
-import { getMonthsAgoDate, toTimeseriesSequenceData } from '../utils/helpers.utils.js'
+import { getMonthsAgoDate } from '../utils/helpers.utils.js'
 import { MetricFactoryFn } from '../services/metrics-producer.service.js'
 import { TenantOnboardingTrendMetric } from '../models/metrics.model.js'
+import { sub } from 'date-fns'
 
 export const getTenantOnboardingTrendMetric: MetricFactoryFn<'statoDiCompletamentoAdesioni'> = (
   _readModel,
@@ -58,4 +59,35 @@ export const getTenantOnboardingTrendMetric: MetricFactoryFn<'statoDiCompletamen
   })
 
   return result
+}
+
+/**
+ * Converts a list of dates into a timeseries sequence data.
+ * @param oldestDate The oldest date in the list, which will be used as the starting point for the timeseries
+ * @param jump The jump between each data point
+ * @param data The list of dates
+ */
+export function toTimeseriesSequenceData({
+  oldestDate,
+  jump,
+  data,
+}: {
+  oldestDate: Date
+  jump: Duration
+  data: Array<Date>
+}): Array<{ date: Date; count: number }> {
+  let currentDate = new Date()
+  let currentCount: number = data.length
+  const timeseriesData: Array<{ date: Date; count: number }> = [{ date: currentDate, count: currentCount }]
+
+  while (oldestDate < currentDate) {
+    // Jump to the next date
+    currentDate = sub(currentDate, jump)
+    // Count the number of dates that are less than or equal to the current date, and add it to the timeseries data
+    currentCount = data.filter((date) => date <= currentDate).length
+
+    timeseriesData.push({ date: currentDate, count: currentCount })
+  }
+  // Reverse the timeseries data so that the oldest date is first
+  return timeseriesData.reverse()
 }
