+++ jobs/dtd-metrics/src/utils/helpers.utils.ts
@@ -2,7 +2,6 @@ import { logInfo, logWarn, logError } from '@interop-be-reports/commons'
 import { randomUUID } from 'crypto'
 import { sub } from 'date-fns'
 import { json2csv as _json2csv } from 'json-2-csv'
-import { Worker } from 'worker_threads'
 
 export function getMonthsAgoDate(numMonths: number): Date {
   const result = sub(new Date(), { months: numMonths })
@@ -58,57 +57,4 @@ export function json2csv(data: object[]): string {
     .join(',')
 
   return [newHeader, ...records].join('\n')
-}
-
-export function splitArrayIntoChunks<T>(array: T[], chunkSize: number): T[][] {
-  if (chunkSize <= 0) throw new Error(`Invalid chunk size: ${chunkSize}`)
-
-  const chunks: T[][] = []
-  for (let i = 0; i < array.length; i += chunkSize) {
-    chunks.push(array.slice(i, i + chunkSize))
-  }
-  return chunks
-}
-
-/**
- * Converts a list of dates into a timeseries sequence data.
- * @param oldestDate The oldest date in the list, which will be used as the starting point for the timeseries
- * @param jump The jump between each data point
- * @param data The list of dates
- */
-export function toTimeseriesSequenceData({
-  oldestDate,
-  jump,
-  data,
-}: {
-  oldestDate: Date
-  jump: Duration
-  data: Array<Date>
-}): Array<{ date: Date; count: number }> {
-  let currentDate = new Date()
-  let currentCount: number = data.length
-  const timeseriesData: Array<{ date: Date; count: number }> = [{ date: currentDate, count: currentCount }]
-
-  while (oldestDate < currentDate) {
-    // Jump to the next date
-    currentDate = sub(currentDate, jump)
-    // Count the number of dates that are less than or equal to the current date, and add it to the timeseries data
-    currentCount = data.filter((date) => date <= currentDate).length
-
-    timeseriesData.push({ date: currentDate, count: currentCount })
-  }
-  // Reverse the timeseries data so that the oldest date is first
-  return timeseriesData.reverse()
-}
-
-export function createWorker<T = unknown>(filename: string, workerData?: unknown): Promise<T> {
-  return new Promise(function (resolve, reject) {
-    const worker = new Worker(filename, { workerData })
-    worker.on('message', (data) => {
-      resolve(data)
-    })
-    worker.on('error', (msg) => {
-      reject(`An error ocurred: ${msg}`)
-    })
-  })
 }
